Initial files from a previous project
This is a refactor of a simple web signup form. I am learning to use
jQuery so that the form works on most browsers. I have set up Webpack
to help with development as it includes a devserver with extra support.

The server displays any errors from the build process, mostly linting
errors.